Migrate to jQuery, disable cookie storage per default
I found that the cookie feature in the older versions wasn't really
useful. The ids changed every time the cache expired and thus the state
didn't persist longer than one day (at most). In this version there are
thus several changes:

 - cookies are now disabled by default but can be enabled again in the
   configuration
 - the ids that are used for identifying the node are now only based on
   the current id and the position in the page. This means that the
   cookie will be valid as long as the position in the page isn't
   changed
 - the outliner plugin now uses its own cookies that expire after a week
   so it won't produce a lot of garbage in the settings cookie
   (previously the values were stored in the settings cookie and were
   never deleted)
 - the old values are removed from the settings cookie

Furthermore the JavaScript code was migrated to jQuery so the plugin
will still work in the next DokuWiki release (but doesn't work anymore
in Rincewind or older).